2024_06_24/requirements.txt
# To install these packages, run the following command in your terminal:
# pip install -r requirements.txt

requests
pydantic
psycopg2-binary